fixed docker issues especially with RandomForest_R (blocking prompt during install) + moved jobs logic to its own module
