+++ frameworks/RandomForest_R/setup.sh
@@ -2,11 +2,11 @@
 HERE=$(dirname "$0")
 . $HERE/../shared/setup.sh
 if [[ -x "$(command -v apt-get)" ]]; then
-    SUDO apt-get -y install software-properties-common apt-transport-https libxml2-dev
+    SUDO apt-get install -y software-properties-common apt-transport-https libxml2-dev
     SUDO apt-key adv --keyserver keyserver.ubuntu.com --recv-keys E084DAB9
     SUDO add-apt-repository 'deb [arch=amd64,i386] https://cran.rstudio.com/bin/linux/ubuntu bionic-cran35/'
     SUDO apt update
-    SUDO apt-get -y install r-base r-base-dev
+    SUDO apt-get install -y r-base r-base-dev
 fi
 #PIP install --no-cache-dir -r $HERE/py_requirements.txt
